* Removed dependancy on SSL library in Windows.

+++ Core/epidatafilestypes.pas
@@ -8,7 +8,7 @@ uses
   variants;
 
 type
-  EpiInteger   = Integer;
+  EpiInteger   = Int64;
   EpiFloat     = Extended;
   EpiDate      = Integer;
   EpiTime      = TDateTime;
